% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildShiny.R
\name{addDendrogramMarkers}
\alias{addDendrogramMarkers}
\title{Add marker genes to reference dendrogram for tree mapping}
\usage{
addDendrogramMarkers(
  dend,
  norm.data,
  metadata,
  celltypeColumn = "cluster_label",
  subsample = 100,
  num.markers = 20,
  de.param = scrattch.hicat::de_param(low.th = 1, padj.th = 0.01, lfc.th = 1, q1.th =
    0.3, q2.th = NULL, q.diff.th = 0.7, de.score.th = 100, min.cells = 2, min.genes = 5),
  calculate.de.genes = TRUE,
  save.shiny.output = TRUE,
  mc.cores = 1,
  bs.num = 100,
  p = 0.7,
  low.th = 0.15,
  shinyFolder = paste0(getwd(), "/")
)
}
\arguments{
\item{dend}{A dendrogram in R format to which marker genes will be added, or a character string with a file location of "dend.RData"}

\item{norm.data}{A matrix of log normalized reference data, or character string with a file location of "data_t.feather".  If a count matrix is provided, the data data will be log normalized.  This should be the data matrix used to generate the dendrogram.}

\item{metadata}{Data frame of metadata with rows corresponding to cells/nuclei, and either row names or a column called "sample_id" corresponding to cell names. This matrix must include entries for all cells in norm.data.  Could also be a file.  Columns can be numeric, categorical, or factors.}

\item{celltypeColumn}{Column name correspond to the cell type names in the dendrogram (default = "cluster_label"). At least two cells per cell type in the dendrogram must be included.}

\item{subsample}{The number of cells to retain per cluster (default = 100)}

\item{num.markers}{The maximum number of markers to calculate per pairwise differential calculation per direction (default = 20)}

\item{de.param}{Differential expression (DE) parameters for genes and clusters used to define marker genes.  By default the values are set to the 10x nuclei defaults from scrattch.hicat, except with min.cells=2 (see the function \code{de_param} in the scrattch.hicat for more details).}

\item{calculate.de.genes}{Default=TRUE. If set to false, the function will search for a file called "de.genes.rda" to load precalculated de genes.}

\item{save.shiny.output}{Should standard output files be generated and saved to the directory (default=TRUE).  These are not required for tree mapping, but are required for building a patch-seq shiny instance.  This is only tested in a UNIX environment.  See notes.}

\item{mc.cores}{Number of cores to use for running this function to speed things up.  Default = 1.  Values>1 are only supported in an UNIX environment and require and \code{foreach} and \code{doParallel} R libraries.}

\item{bs.num, p, low.th}{Extra variables for the \code{map_dend_membership} function in scrattch.hicat.  Defaults are set reasonably.}

\item{shinyFolder}{The location to save shiny output, if desired

NOTES

If save.shiny.output=TRUE, the following files will be generated:
reference.rda, which includes a variable \code{reference} as follows:
reference$cl.dat - These are the cluster means that are used for mapping comparisons
reference$dend   - This is the dendrogram with marker genes attached
membership_information_reference.rda, which includes two variables
\code{memb.ref}   - matrix indicating how much confusion there is the mapping between each cell all of the nodes in the tree (including all cell types) when comparing clustering and mapping results with various subsamplings of the data
\code{map.df.ref} - Result of tree mapping for each cell in the reference against the clustering tree, including various statistics and marker gene evidence.  This is the same output that comes from tree mapping.#'}
}
\value{
An updated dendrogram variable that is the same as \code{dend} except with marker genes added to each node.
}
\description{
Add marker genes to reference dendrogram for tree mapping
}
