% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchseqQC.R
\name{applyPatchseqQC}
\alias{applyPatchseqQC}
\title{This function applies patchseqQC, given a taxonomy and query data}
\usage{
applyPatchseqQC(AIT.anndata, query.data, query.metadata, verbose = FALSE)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy object.}

\item{query.data}{A count or CPM matrix for the query data.}

\item{query.metadata}{A data frame of metadata for the query data.}

\item{verbose}{Should status be printed to the screen?}
}
\value{
A new query.metadata file with appended QC columns
}
\description{
This function applies patchseqQC, given a taxonomy and query data
}
