% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{batch_process}
\alias{batch_process}
\title{Batch process}
\usage{
batch_process(x, batch.size, FUN, mc.cores = 1, .combine = "c", ...)
}
\arguments{
\item{x}{to_be_added}

\item{batch.size}{to_be_added}

\item{FUN}{to_be_added}

\item{mc.cores}{to_be_added}

\item{.combine}{to_be_added}

\item{...}{}
}
\value{
to_be_added
}
\description{
Batch process
}
\keyword{internal}
