% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_dend}
\alias{build_dend}
\title{Build dend (updated to specify dendextend version of "set")}
\usage{
build_dend(
  cl.dat,
  cl.cor = NULL,
  l.rank = NULL,
  l.color = NULL,
  nboot = 100,
  ncores = 1
)
}
\arguments{
\item{cl.dat}{Normalized data of the REFERENCE data set}

\item{cl.cor}{Matrix of cell x cell correlations (calculated if not provided)}

\item{l.rank}{Factor of cluster order (in a specific format)}

\item{l.color}{Factor of clluster colors (in a specific format)}

\item{nboot}{Number of bootstrapping runs to calculate the membership from (default = 100)}

\item{ncores}{Number of cores for performing calculations}
}
\value{
dendrogram and a couple of related things
}
\description{
Build dend (updated to specify dendextend version of "set")
}
\keyword{internal}
