% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{build_marker_index_node_cl}
\alias{build_marker_index_node_cl}
\title{INFO -- PLEASE ADD --}
\usage{
build_marker_index_node_cl(
  lvl,
  nlevel,
  nodestr,
  cl.df,
  marker_index,
  children,
  pre.marker_index = NA,
  query.genes,
  train.dat,
  mc.cores = 10,
  div_thr = 5,
  de.dir,
  subsample_pct,
  top.n.genes = 15,
  n.group.genes = 3000,
  select.markers,
  cl.bin,
  outdir
)
}
\arguments{
\item{cl.df}{to_be_added}

\item{in.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
