% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{build_marker_index_tree_cl}
\alias{build_marker_index_tree_cl}
\title{INFO -- PLEASE ADD --}
\usage{
build_marker_index_tree_cl(
  train.list,
  pre.train.list = NA,
  query.genes = NA,
  outdir,
  div_thr = 3,
  subsample_pct = 0.9,
  top.n.genes = 15,
  n.group.genes = 3000,
  mc.cores = 10
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
