% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_reference}
\alias{build_reference}
\title{Function for building the standard reference format, including adding marker genes to the clustering tree}
\usage{
build_reference(
  cl,
  norm.dat,
  dend,
  de.genes,
  cl.label = NULL,
  up.gene.score = NULL,
  down.gene.score = NULL,
  n.markers = 30
)
}
\arguments{
\item{cl}{Factor vector where values are cluster ids (e.g., a numeric vector of corresponding to cell type order in the tree) and values are sample ids for cells (e.g., this vector has length = number of cells)}

\item{norm.dat}{log normalized expression data}

\item{dend}{Input dendrogram}

\item{de.genes}{output from \code{display_cl} function}

\item{cl.label}{Factor vector here values are cluster ids (e.g., a numeric vector of corresponding to cell type order in the tree) and values are dendrogram labels (e.g., this vector has length = number of clusters)}

\item{up.gene.score}{Output from \code{get_gene_score}}

\item{down.gene.score}{Output from \code{get_gene_score}}

\item{n.markers}{Number of marker genes to return per comparison (default=30)}
}
\value{
A list where \code{dend} is the updated dendrogram with markers attached and \code{cl.dat} is a matrix of cluster means
}
\description{
Function for building the standard reference format, including adding marker genes to the clustering tree
}
\keyword{internal}
