% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_utils.R
\name{build_train_index_bs}
\alias{build_train_index_bs}
\title{INFO -- PLEASE ADD --}
\usage{
build_train_index_bs(
  cl.dat,
  method = c("Annoy.Cosine", "cor", "Annoy.Euclidean"),
  sample.markers.prop = 0.8,
  iter = 100,
  mc.cores = 10,
  fn = tempfile(fileext = ".idx")
)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
