% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN_build.R
\name{build_train_list_default}
\alias{build_train_list_default}
\title{build_train_list_default : building train.list}
\usage{
build_train_list_default(
  pre.train.list = NA,
  query.genes = NA,
  TrainDir,
  TaxDir,
  prefix = "",
  nlevel = 4,
  TaxFN,
  rm.cl = NULL,
  root.markers = NULL
)
}
\arguments{
\item{pre.train.list}{to_be_added}

\item{query.genes}{to_be_added}

\item{TaxDir}{to_be_added}

\item{prefix}{to_be_added}

\item{nlevel}{to_be_added}

\item{TaxFN}{to_be_added}

\item{rm.cl}{to_be_added}

\item{root.markers}{to_be_added}
}
\value{
train.list
}
\description{
build_train_list_default : building train.list
}
