% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN_build.R
\name{build_train_list_FBLO10}
\alias{build_train_list_FBLO10}
\title{INFO -- PLEASE ADD --}
\usage{
build_train_list_FBLO10(
  pre.train.list = NA,
  query.genes,
  CLSDIR,
  TaxDir,
  prefix = "",
  nlevel = 4,
  TaxFN,
  rm.cl = c()
)
}
\arguments{
\item{x}{to_be_added}

\item{key}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
