% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized.R
\name{build_train_list_on_taxonomy}
\alias{build_train_list_on_taxonomy}
\title{Starting point for optimized tree mapping}
\usage{
build_train_list_on_taxonomy(
  TaxFN = NA,
  Taxonomy,
  pre.train.list = NA,
  query.genes = NA,
  prefix = "",
  mapping.method = c("flat", "hierarchy"),
  prebuild = FALSE,
  newbuild = FALSE,
  mc.cores = 10,
  div_thr = 3,
  subsample_pct = 0.9,
  top.n.genes = 15,
  n.group.genes = 3000,
  rm.cl = c()
)
}
\arguments{
\item{rm.cl}{}
}
\value{
Mapping results
}
\description{
Starting point for optimized tree mapping
}
