% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_build.R
\name{build_train_list_TH16_2}
\alias{build_train_list_TH16_2}
\title{INFO -- PLEASE ADD --}
\usage{
build_train_list_TH16_2(
  pre.train.list = NA,
  query.genes = NA,
  WBDIR,
  TaxDir,
  prefix = "",
  nlevel = 4,
  TaxFN,
  rm.cl = c()
)
}
\arguments{
\item{key}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
