% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNBuild.R
\name{build_train_list_WB17_FB}
\alias{build_train_list_WB17_FB}
\title{INFO -- PLEASE ADD --}
\usage{
build_train_list_WB17_FB(
  pre.train.list = NA,
  query.genes = NA,
  WBDIR,
  TaxDir,
  prefix = "",
  nlevel = 4,
  TaxFN,
  rm.cl = c()
)

build_train_list_WB17_FB(
  pre.train.list = NA,
  query.genes = NA,
  WBDIR,
  TaxDir,
  prefix = "",
  nlevel = 4,
  TaxFN,
  rm.cl = c()
)

build_train_list_WB17_FB(
  pre.train.list = NA,
  query.genes = NA,
  WBDIR,
  TaxDir,
  prefix = "",
  nlevel = 4,
  TaxFN,
  rm.cl = c()
)
}
\arguments{
\item{x}{to_be_added}

\item{key}{to_be_added}
}
\value{
???

???

???
}
\description{
INFO -- PLEASE ADD --

INFO -- PLEASE ADD --

INFO -- PLEASE ADD --
}
\keyword{internal}
