% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPatchseqTaxonomy.R
\name{buildPatchseqTaxonomy}
\alias{buildPatchseqTaxonomy}
\title{Save marker genes for patchSeqQC}
\usage{
buildPatchseqTaxonomy(
  AIT.anndata,
  mode.name = "patchseq",
  subsample = 100,
  subclass.column = "subclass_label",
  class.column = "class_label",
  off.target.types = c("Glia", "glia", "non-neuronal", "Non-neuronal"),
  num.markers = 50,
  taxonomyDir = AIT.anndata$uns$taxonomyDir
)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}

\item{mode.name}{A name to identify the new taxonomy version.}

\item{subsample}{The number of cells to retain per cluster (default = 100).}

\item{subclass.column}{Column name corresponding to the moderate-resolution cell types used for the cell types of interest (default = "subclass_label").}

\item{class.column}{Column name corresponding to the low-resolution cell types used for the off-target cell types (default = "class_label").}

\item{off.target.types}{A character vector of off-target (also known as 'contamination') cell types.  This must include at least one of the cell types found in "class.column" and/or "subclass.column" (both columns are checked)}

\item{num.markers}{The maximum number of markers to calculate per node per direction (default = 50)}

\item{taxonomyDir}{= The location to save shiny output (default = current working directory).

Nothing is returned; however an R data object called "QC_markers.RData" is returned with the following variables
markers,
countsQC,
cpmQC,
classBr,
subclassF,
allMarkers,}
}
\value{
AIT.anndata
}
\description{
This function write a file called \code{QC_markers.RData} that contains all the variables required for applying the patchseq QC algorithm \code{pathseqtools} (which is an more flexible version of the \code{patchSeqQC} algorithm).  This is only used for patch-seq analysis.
}
