% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildShiny.R
\name{buildReference}
\alias{buildReference}
\title{This function builds the minimum files required for Shiny}
\usage{
buildReference(
  counts,
  meta.data,
  feature.set,
  umap.coords,
  shinyFolder,
  cluster_colors = NULL,
  metadata_names = NULL,
  subsample = 2000,
  reorder.dendrogram = FALSE
)
}
\arguments{
\item{counts}{A count matrix in sparse format: dgCMatrix.}

\item{meta.data}{Meta.data corresponding to count matrix. Rownames must be equal to colnames of counts.}

\item{feature.set}{Set of feature used to calculate dendrogram. Typically highly variable and/or marker genes.}

\item{umap.coords}{Dimensionality reduction coordiant data.frame with 2 columns. Rownames must be equal to colnames of counts.}

\item{shinyFolder}{The location to save Shiny objects, e.g. "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/10x_seq/NHP_BG_20220104/"}

\item{cluster_colors}{An optional named character vector where the values correspond to colors and the names correspond to celltypes in celltypeColumn.  If this vector is incomplete, a warning is thrown and it is ignored.}

\item{metadata_names}{An optional named character vector where the vector NAMES correspond to columns in the metadata matrix and the vector VALUES correspond to how these metadata should be displayed in Shiny. This is used for writing the desc.feather file later.}

\item{subsample}{The number of cells to retain per cluster}

\item{reorder.dendrogram}{Should dendogram attempt to match a preset order? (Default = FALSE).  If TRUE, the dendrogram attempts to match the celltype factor order as closely as possible (if celltype is a character vector rather than a factor, this will sort clusters alphabetically, which is not ideal).}
}
\value{

}
\description{
This function builds the minimum files required for Shiny
}
