% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildShiny.R
\name{buildReferenceFolder}
\alias{buildReferenceFolder}
\title{Starting from a Seurat object this function builds the minimum files required for Shiny}
\usage{
buildReferenceFolder(
  seurat.obj,
  shinyFolder,
  subsample = 2000,
  feature.set = NULL,
  reorder.dendrogram = FALSE,
  save.normalized.data = FALSE
)
}
\arguments{
\item{seurat.obj}{A Seurat object as specified in the notes, or as output from \code{createSeuratObjectForReferenceFolder}}

\item{shinyFolder}{The location to save Shiny objects, e.g. "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/10x_seq/NHP_BG_20220104/"}

\item{subsample}{The number of cells to retain per cluster}

\item{feature.set}{An optional user supplied gene set otherwise Seurat obects variable features are used}

\item{reorder.dendrogram}{Should dendogram attempt to match a preset order? (Default = FALSE).  If TRUE, the dendrogram attempts to match the celltype factor order as closely as possible (if celltype is a character vector rather than a factor, this will sort clusters alphabetically, which is not ideal).}

\item{save.normalized.data}{Should normalized data (TRUE="data" slot in Seurat object) or raw count matrix (FALSE="counts" slot in Seurat object; default) be saved?}
}
\value{

}
\description{
Starting from a Seurat object this function builds the minimum files required for Shiny
}
