% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized.R
\name{call_ANN_0}
\alias{call_ANN_0}
\title{INFO -- PLEASE ADD --}
\usage{
call_ANN_0(index.bs, qdat, prev, children, gg)
}
\arguments{
\item{children}{}

\item{gg}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
