% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{call_membership_cl}
\alias{call_membership_cl}
\title{INFO -- PLEASE ADD --}
\usage{
call_membership_cl(
  lvl,
  index.bs,
  qdat,
  membership.df,
  ancestor,
  blocksize = 50000,
  mc.cores = 10,
  topk = 3
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
