% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{check_pairs_ds}
\alias{check_pairs_ds}
\title{INFO -- PLEASE ADD --}
\usage{
check_pairs_ds(
  de.dir,
  to.add,
  genes,
  cl.bin,
  de = NULL,
  mc.cores = 10,
  max.num = 1000
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
