% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN.R
\name{check_reference_ready}
\alias{check_reference_ready}
\title{Check h5adFN and update adata}
\usage{
check_reference_ready(h5adFN, flag.root = TRUE)
}
\arguments{
\item{h5adFN}{: output of prepareTaxonomy()}

\item{flag.root}{: at root node of HANN, use mfish 500 marker genes}
}
\value{
adata : anndata updated
}
\description{
Check h5adFN and update adata
}
