% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingComparisons.R
\name{compactness_distance}
\alias{compactness_distance}
\title{Calculate a compactness score}
\usage{
compactness_distance(
  query.data,
  query.group,
  query.secondary = NULL,
  variable.features = rownames(query.data)
)
}
\arguments{
\item{query.data}{A logCPM normalized matrix of the data}

\item{query.group}{A group vector to calculate compactness distance over (e.g., cluster assignments)}

\item{query.secondary}{An optional secondary group vector for comparison with the primary group vector (e.g., cortical layer or transgenic line)}

\item{variable.features}{A precomputed set of variable features.  If not provided, all genes are used.}
}
\value{
A vector of compactness scores for each cell
}
\description{
This function calculates the compactness score, defined as the the average (Pearson) correlation-based distance between each cell and the assigned group centroid (median) using the variable genes.  If a secondary group is provided (e.g., transgenic line, cortical layer, etc.), the function first sets gene expression values for each cell as expression values for the cluster median and then returns compactness per cell summarized by the secondary group.
}
\keyword{internal}
