% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingComparisons.R
\name{compare_heatmap}
\alias{compare_heatmap}
\title{Compare and plot two sets of cluster assignments as a heatmap}
\usage{
compare_heatmap(
  cl,
  ref.cl,
  threshold = 0.2,
  cexLab = NULL,
  Rowv = NA,
  Colv = NA,
  ylab = NULL,
  xlab = NULL,
  main = NULL,
  margins = c(6, 6),
  scale = "none",
  trace = "none",
  dendrogram = "none",
  ...
)
}
\arguments{
\item{cl}{A cluster factor object to compare to a reference}

\item{ref.cl}{A cluster factor object for the reference clusters}

\item{threshold}{Maximum value to show in heatmap.  Lower values will highlight off-target expression more.}

\item{cexLab}{Size of the label names to display on the screen.  The function will attempt to guess this if not inputted, adjustments may be needed if not all labels are shown.}

\item{...}{Other inputs to the function \code{heatmap.3}}
}
\value{
a list with output from heatmap.3, after displaying the heatmap to the screen.
}
\description{
This is a wrapper for the function \code{heatmap.3} in scrattch.hicat
}
\keyword{internal}
