% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrMap.R
\name{cor}
\alias{cor}
\title{Correlation based cell type mapping wrapper}
\usage{
cor(...)
}
\arguments{
\item{nGenes}{}
}
\value{
Correlation mapping results
}
\description{
Correlation based cell type mapping wrapper
}
\keyword{internal}
