% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN_utils.R
\name{create_pairs}
\alias{create_pairs}
\title{Title}
\usage{
create_pairs(
  cn1,
  cn2 = cn1,
  direction = "nondirectional",
  include.self = FALSE
)
}
\arguments{
\item{direction}{DESCRIPTION}

\item{include.self}{DESCRIPTION}

\item{cn}{DESCRIPTION}
}
\value{
???
}
\description{
Title
}
\keyword{internal}
