% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildShiny.R
\name{createSeuratObjectForReferenceFolder}
\alias{createSeuratObjectForReferenceFolder}
\title{Build a Seurat object for buildReferenceFolder}
\usage{
createSeuratObjectForReferenceFolder(
  counts,
  metadata,
  clusterColumn = "cluster",
  celltypeColumn = clusterColumn,
  normalizebyLogCPM = TRUE,
  generateUMAP = TRUE,
  npcs = 10,
  feature.set = NULL,
  celltype_colors = NULL,
  metadata_names = setNames(colnames(metadata), colnames(metadata)),
  subsample = Inf,
  gene_names = rownames(counts),
  cell_names = colnames(counts)
)
}
\arguments{
\item{counts}{Count matrix with genes as rows and cells/nuclei as columns}

\item{metadata}{Matrix of metadata with rows corresponding to cells/nuclei, and row names corresponding to cell names. IF column names of counts are unspecified, then this matrix must be the same length and ordered in the same way as the counts matrix.  In any case, this must be a superset of counts (e.g., no missing metadata)}

\item{clusterColumn}{Column name correspond to the clusters to include in the clustering results}

\item{celltypeColumn}{Column name correspond to the cell type names to include in the dendrogram.  By default (and in most cases) this is identical to clusterColumn.}

\item{normalizebyLogCPM}{If TRUE (default) log(CPM+1) is used for normalization. If FALSE, \code{NormalizeData} Seurat function is used}

\item{generateUMAP}{A boolean (default=TRUE) indicating whether UMAP should be calculated}

\item{npcs}{Number of pcs used for building the UMAP (default = 10)}

\item{feature.set}{An optional user supplied gene set otherwise Seurat objects variable features are used for umap generation}

\item{celltype_colors}{An optional named character vector where the values correspond to colors and the names correspond to celltypes in celltypeColumn.  If this vector is incomplete, a warning is thrown and it is ignored.}

\item{metadata_names}{An optional named character vector where the vector NAMES correspond to columns in the metadata matrix and the vector VALUES correspond to how these metadata should be displayed in Shiny. This is used for writing the desc.feather file later.}

\item{subsample}{The number of cells to retain per cluster (default is to keep all of them)}

\item{gene_names}{Gene names corresponding to rows in the count matrix (by default, checks the rownames)}

\item{cell_names}{Sample names corresponding to the columns in the count matrix (by default, checks the column names)}
}
\value{

}
\description{
This function builds a Seurat object from a count matrix and metadata table for use with buildReferenceFolder, along with some other optional inputs. Unless otherwise specific log2(CPM(counts)+1) are saved in the "data" slot of the Seurat object because that is what is used for most downstream analysis.
}
