% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.convert_gene_input_to_vector}
\alias{.convert_gene_input_to_vector}
\title{Convert entered gene set to a logical vector, or return errors}
\usage{
.convert_gene_input_to_vector(AIT.anndata, genes.to.use)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy object.}

\item{genes.to.use}{The set of genes to use for correlation calculation and/or Seurat integration (default is the highly_variable_genes associated with the current mode). Can be (1) a character vector of gene names, (2) a TRUE/FALSE (logical) vector of which genes to include, or (3) a column name in AIT.anndata$var corresponding to a logical vector of variable genes.}
}
\value{
entered gene set as a logical vector
}
\description{
Convert entered gene set to a logical vector, or return errors
}
\keyword{internal}
