% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file.path}
\alias{file.path}
\title{Function to set mapping mode}
\usage{
file.path(
  ...,
  path_separator = getOption("path_separator"),
  leading_string = getOption("leading_string"),
  change_chars = NULL
)

file.path(
  ...,
  path_separator = getOption("path_separator"),
  leading_string = getOption("leading_string"),
  change_chars = NULL
)
}
\arguments{
\item{...}{character vectors. Long vectors are not supported.}

\item{path_separator}{the path separator to use (assumed to be ASCII).}

\item{leading_string}{what is the leading character(s) (e.g., '/' or '\\\\'). A string can be provided, or by default if the "leading_string" global variable is set it takes that variable, otherwise this is guessed at based on operating system and/or existence of a file at the file path}

\item{change_chars}{which characters should be changes to the path_separator value (default NULL = "none"). If you want to change all slashes in the file path to the correct direction set change_chars = c("/","\\"))}

\item{AIT.anndata}{A vector of cluster names in the reference taxonomy.}

\item{mode}{Number of cells to keep per cluster.}
}
\value{
AIT anndata with mode set for mapping

A file path with slashes going the correct direction.
}
\description{
Construct the path to a file from components in a platform-independent way. This version is a wrapper of the base \code{file.path} function which also reverses '/' direction and can attempt to add double slashes if needed.
}
