% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalMapMyCells.R
\name{get_anndata_path}
\alias{get_anndata_path}
\title{This function saves the AIT reference taxonomy to a temp folder as h5ad, if the provided file path is invalid.}
\usage{
get_anndata_path(anndata_path, temp_folder)
}
\arguments{
\item{anndata_path}{Local file path of the AIT reference taxonomy (h5ad file).}
}
\value{
Local file path to the AIT reference taxonomy h5ad file.
}
\description{
This function saves the AIT reference taxonomy to a temp folder as h5ad, if the provided file path is invalid.
}
\keyword{internal}
