% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_utils.R
\name{get_de_genes}
\alias{get_de_genes}
\title{INFO -- PLEASE ADD --}
\usage{
get_de_genes(ds, cl.bin, cl1, cl2)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
