% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalMapMyCells.R
\name{get_hierarchical_extended_results}
\alias{get_hierarchical_extended_results}
\title{This function returns names and bootstrap probabilities from all top mapped cell sets}
\usage{
get_hierarchical_extended_results(extended_result_path)
}
\arguments{
\item{extended_result_path}{Full file path and name where the original mapping results will be saved.}
}
\value{
A table of the top cell set names and bootstrap probabilities from top results and runner-up results for each level of the hierarchy included in the extended_result_path file.  By default, this is the top 5 most likely cell sets for either (1) multiple hierachy levels in hierarchical mapping or (2) a single hierarchy level for correlation mapping.
}
\description{
This function returns names and bootstrap probabilities from all top mapped cell sets
}
