% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_utils.R
\name{get_knn}
\alias{get_knn}
\title{Get KNN}
\usage{
get_knn(
  dat,
  ref.dat,
  k,
  method = "cor",
  dim = NULL,
  index = NULL,
  build.index = FALSE,
  transposed = TRUE,
  return.distance = FALSE
)
}
\arguments{
\item{dat}{}

\item{ref.dat}{}

\item{k}{}

\item{method}{}

\item{dim}{}
}
\value{

}
\description{
Get KNN
}
\keyword{internal}
