% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{get_knn_batch}
\alias{get_knn_batch}
\title{get knn batch}
\usage{
get_knn_batch(
  dat,
  ref.dat,
  k = 1,
  method = "cor",
  dim = NULL,
  batch.size,
  mc.cores = 1,
  return.distance = FALSE,
  ...
)
}
\arguments{
\item{dat}{to_be_added}

\item{ref.dat}{to_be_added}

\item{k}{to_be_added}

\item{method}{to_be_added}

\item{dim}{to_be_added}

\item{batch.size}{to_be_added}

\item{mc.cores}{to_be_added}
}
\value{
to_be_added
}
\description{
get knn batch
}
\keyword{internal}
