% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized.R
\name{get_last_field}
\alias{get_last_field}
\title{INFO -- PLEASE ADD --}
\usage{
get_last_field(x, key)
}
\arguments{
\item{key}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
