% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalMapMyCells.R
\name{get_mapmycells_results}
\alias{get_mapmycells_results}
\title{This function maps query data against the AIT (Shiny) reference taxonomy using hierarchical mapping.}
\usage{
get_mapmycells_results(
  query_data_output_path,
  extended_result_path,
  precomp_stats_output_path,
  query_markers_output_path,
  normalization,
  chunk_size,
  n_processors
)
}
\arguments{
\item{query_data_output_path}{Full file path to the query data, a logCPM normalized matrix to be annotated that's saved as AnnData H5AD file.}

\item{extended_result_path}{Full file path and name where the original mapping results will be saved.}

\item{precomp_stats_output_path}{Full file path to the precompute stats HDF5 file.}

\item{query_markers_output_path}{Full file path to the query markers JSON file.}

\item{normalization}{Normalization of the h5ad files; must be either 'raw' or 'log2CPM'.}

\item{chunk_size}{Number of rows each worker process should load at a time from the query dataset.}

\item{n_processors}{Number of independent worker processes to spin up.}
}
\value{
resuls
}
\description{
This function maps query data against the AIT (Shiny) reference taxonomy using hierarchical mapping.
}
\keyword{internal}
