% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalMapMyCells.R
\name{get_marker_genes}
\alias{get_marker_genes}
\title{Retrieves the query markers from the AIT taxonomy and saves it in JSON format. The markers are, for every pair of leaf nodes in the cell type taxonomy tree,
every gene that could conceivably be a marker gene for discriminating between those two cell types.}
\usage{
get_marker_genes(AIT_anndata, query_markers_output_path, temp_folder)
}
\arguments{
\item{AIT_anndata}{A reference taxonomy anndata object.}

\item{query_markers_output_path}{Alternative path to the user provided query markers JSON file. Will be generated, if not provided.}

\item{temp_folder}{Temporary directory for writing out the hierarchical files (the code will clean these up after itself).}
}
\value{
File path of the query markers file.
}
\description{
Retrieves the query markers from the AIT taxonomy and saves it in JSON format. The markers are, for every pair of leaf nodes in the cell type taxonomy tree,
every gene that could conceivably be a marker gene for discriminating between those two cell types.
}
\keyword{internal}
