% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Detect the operating system}
\usage{
get_os()
}
\description{
This function was taken directly from https://conjugateprior.org/2015/06/identifying-the-os-from-r/ and all credit goes to Will Lowe from "conjugateprior".
}
\keyword{internal}
