% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalMapMyCells.R
\name{get_precomp_stats}
\alias{get_precomp_stats}
\title{Retrieves the precompute stats data from the AIT taxonomy and saves it in H5AD format. precomputed_stats.h5 file contains a serialization of the cell type taxonomy tree,
as well as statistics about the reference cells that have been assigned to the cell types in the taxonomy.}
\usage{
get_precomp_stats(
  AIT_anndata,
  anndata_path,
  precomp_stats_output_path,
  temp_folder
)
}
\arguments{
\item{AIT_anndata}{A reference taxonomy anndata object.}

\item{anndata_path}{Local file path of the AIT reference taxonomy (h5ad file).}

\item{precomp_stats_output_path}{Alternative path to the user provided precompute stats HDF5 file. Will be generated, if not provided.}

\item{temp_folder}{Temporary directory for writing out the hierarchical files (the code will clean these up after itself).}
}
\value{
File path to the precompute stats file.
}
\description{
Retrieves the precompute stats data from the AIT taxonomy and saves it in H5AD format. precomputed_stats.h5 file contains a serialization of the cell type taxonomy tree,
as well as statistics about the reference cells that have been assigned to the cell types in the taxonomy.
}
\keyword{internal}
