% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalMapMyCells.R
\name{get_query_data_path}
\alias{get_query_data_path}
\title{Saves the query data count matrix as AnnData H5AD file. This file path is later passed into the hierarchical mapping algorithm.}
\usage{
get_query_data_path(query_data, temp_folder)
}
\arguments{
\item{query_data}{A logCPM normalized matrix to be annotated.}

\item{temp_folder}{Temporary directory for writing out the hierarchical files (the code will clean these up after itself).}
}
\value{
File path of the query anndata file.
}
\description{
Saves the query data count matrix as AnnData H5AD file. This file path is later passed into the hierarchical mapping algorithm.
}
\keyword{internal}
