% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingClass.R
\name{getMappingResults}
\alias{getMappingResults}
\title{Get cell type annotations}
\usage{
getMappingResults(x)
}
\value{
mapping results as a data.frame with labels in map.Method and scores in score.Method
}
\description{
Extract cell type annotations from mappingClass S4 class
}
