% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalMapMyCells.R
\name{hierarchicalMapMyCells}
\alias{hierarchicalMapMyCells}
\title{This function maps query data against the AIT (Shiny) reference taxonomy using hierarchical mapping.}
\usage{
hierarchicalMapMyCells(
  AIT_anndata,
  query_data,
  mapping_params_list = list(),
  flat_mapping = FALSE,
  tmp_dir = NULL,
  user_extended_result_path = NULL,
  user_precomp_stats_path = NULL,
  user_query_markers_path = NULL
)
}
\arguments{
\item{AIT_anndata}{A reference taxonomy anndata object.}

\item{query_data}{A logCPM normalized matrix to be annotated.}

\item{tmp_dir}{Temporary directory for writing out the hierarchical files.}

\item{user_extended_result_path}{Full file path and name where the original mapping results will be saved.}

\item{user_precomp_stats_path}{Alternative path to the user provided precompute stats HDF5 file. Will be generated, if not provided.}

\item{user_query_markers_path}{Alternative path to the user provided query markers JSON file. Will be generated, if not provided.

Note: this hierarchical mapping is a wrapper around cell_type_mapper,
and call's it's functions to generate needed files needed for mapping.}

\item{chunk_size}{Number of rows each worker process should load at a time from the query dataset.}

\item{n_processors}{Number of independent worker processes to spin up.}

\item{normalization}{Normalization of the h5ad files; must be either 'raw' or 'log2CPM'.}
}
\value{
List of mapping results with labels and scores.
}
\description{
This function maps query data against the AIT (Shiny) reference taxonomy using hierarchical mapping.
}
