% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{l2norm}
\alias{l2norm}
\title{INFO -- PLEASE ADD --}
\usage{
l2norm(X, by = "column")
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
