% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGEXRef.R
\name{loadGEXRef}
\alias{loadGEXRef}
\title{Read in a reference data set in Allen taxonomy format}
\usage{
loadGEXRef(
 
    refFolder = "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/10x_seq/human_BGplus_20211020",
  sample_field = "sample_id",
  nGenes = 2000,
  hGenes = NULL,
  sub.sample = 1000
)
}
\arguments{
\item{refFolder}{Directory containing the Shiny taxonomy.}

\item{nGenes}{Number of variable genes to compute.}

\item{hGenes}{User supplied variable gene vector.}

\item{sub.sample}{Number of cells to keep per cluster.}

\item{sample_name_field}{Field in reference taxonomy that defines the sample_id.}
}
\value{
Organized reference object ready for mapping against.
}
\description{
Read in a reference data set in Allen taxonomy format
}
