% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxonomy.R
\name{loadTaxonomy}
\alias{loadTaxonomy}
\title{Read in a reference data set in Allen taxonomy format}
\usage{
loadTaxonomy(
  refFolder,
  sample_id = "sample_id",
  nGenes = 2000,
  hGenes = NULL,
  sub.sample = 1000
)
}
\arguments{
\item{refFolder}{Directory containing the Shiny taxonomy.}

\item{sample_id}{Field in reference taxonomy that defines the sample_id.}

\item{nGenes}{Number of variable genes to compute.}

\item{hGenes}{User supplied variable gene vector.}

\item{sub.sample}{Number of cells to keep per cluster.}
}
\value{
Organized reference object ready for mapping against.
}
\description{
Read in a reference data set in Allen taxonomy format
}
