% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{look_up_ancestor}
\alias{look_up_ancestor}
\title{INFO -- PLEASE ADD --}
\usage{
look_up_ancestor(cl.df, nlvl = 3)
}
\arguments{
\item{cl.df}{to_be_added}

\item{in.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
