% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized.R
\name{look_up_children}
\alias{look_up_children}
\title{INFO -- PLEASE ADD --}
\usage{
look_up_children(cl.df, nlvl = 3)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
