% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_utils.R
\name{map_cells_knn_big}
\alias{map_cells_knn_big}
\title{INFO -- PLEASE ADD --}
\usage{
map_cells_knn_big(
  big.dat,
  cl.dat,
  select.cells,
  train.index = NULL,
  method = c("Annoy.Cosine", "cor"),
  batch.size = 10000,
  mc.cores = 10
)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
