% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_utils.R
\name{map_cells_knn_bs}
\alias{map_cells_knn_bs}
\title{INFO -- PLEASE ADD --}
\usage{
map_cells_knn_bs(
  topk = 1,
  test.dat,
  iter = 100,
  cl.dat = NULL,
  train.index.bs = NULL,
  method = c("Annoy.Cosine", "cor"),
  mc.cores = 20,
  ...
)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
