% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_dend}
\alias{map_dend}
\title{Title}
\usage{
map_dend(
  dend,
  cl.dat,
  map.dat,
  select.cells = colnames(map.dat),
  p = 0.8,
  low.th = 0.2,
  default.markers = NULL,
  seed = 42
)
}
\arguments{
\item{dend}{}

\item{cl.dat}{}

\item{map.dat}{}

\item{select.cells}{}

\item{p}{}

\item{low.th}{}

\item{default.markers}{}

\item{seed}{= random seed}

\item{cl}{}

\item{dat}{}
}
\value{
tree mapping to the dendrogram table (cells x nodes with values as probabilities)
}
\description{
Title
}
