% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_dend_membership}
\alias{map_dend_membership}
\title{Title}
\usage{
map_dend_membership(
  dend,
  cl.dat,
  map.dat,
  map.cells,
  mc.cores = 10,
  bs.num = 100,
  seed = 42,
  ...
)
}
\arguments{
\item{dend}{}

\item{cl.dat}{}

\item{map.dat}{}

\item{map.cells}{}

\item{mc.cores}{}

\item{bs.num}{}

\item{seed}{= random seed}

\item{...}{}

\item{cl}{}

\item{dat}{}
}
\value{
membership table
}
\description{
Title
}
