% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized.R
\name{mapping_by_block_parallel_tmp}
\alias{mapping_by_block_parallel_tmp}
\title{INFO -- PLEASE ADD --}
\usage{
mapping_by_block_parallel_tmp(
  query,
  train.list,
  blocksize = 10000,
  mc.cores = 7,
  method = "cor",
  flag.fuzzy = FALSE,
  topk = 1,
  tmpdir = paste0("./mapping_tmp_", Sys.Date())
)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
