% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized.R
\name{mapping_by_block_serial}
\alias{mapping_by_block_serial}
\title{INFO -- PLEASE ADD --}
\usage{
mapping_by_block_serial(
  query.dat,
  train.list,
  blocksize = 10000,
  mc.cores = 7,
  method = "cor",
  iter,
  topk = 1,
  flag.fuzzy = FALSE
)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
