% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingClass.R
\docType{class}
\name{mappingClass-class}
\alias{mappingClass-class}
\title{mappingClass S4 class definition}
\description{
Define the mappingClass that stores mapping results and detailed method outputs from scrattch.mapping
}
