% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingClass.R
\name{mappingClass}
\alias{mappingClass}
\title{Constructor function for mappingClass S4 class}
\usage{
mappingClass(annotations, detailed_results)
}
\arguments{
\item{annotations}{A reference taxonomy anndata object.}

\item{detailed_results}{a method-specific set of additional mapping output}
}
\value{
Instance of mappingClass S4 class.
}
\description{
This function instantiates a mappingClass S4 class object.
}
\examples{
resultAnno <- mappingClass(
  annotations = data.frame(
    map.Corr = c("Exc", "Inh", "Inh"),
    score.Corr = c(0.9, 0.9, 0.9)
  ),
  detailed_results = list(
    corr = NA,
    tree = membership, ## For patchseq directory
    seurat = NA
  )
)

}
