% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mappingMode}
\alias{mappingMode}
\title{Function to set mapping mode}
\usage{
mappingMode(AIT.anndata, mode)
}
\arguments{
\item{AIT.anndata}{A vector of cluster names in the reference taxonomy.}

\item{mode}{Number of cells to keep per cluster.}
}
\value{
AIT anndata with mode set for mapping
}
\description{
Function to set mapping mode
}
