% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{predict_HKNN_cl}
\alias{predict_HKNN_cl}
\title{INFO -- PLEASE ADD --}
\usage{
predict_HKNN_cl(
  lvl,
  iter_i,
  marker_index,
  query,
  assigned.df,
  children,
  ancestor,
  nodestr = "",
  topk
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
