% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{predict_HKNN_cl_bs}
\alias{predict_HKNN_cl_bs}
\title{INFO -- PLEASE ADD --}
\usage{
predict_HKNN_cl_bs(
  query,
  train.list,
  marker_index,
  iter = 100,
  mc.cores = 5,
  method = "cor",
  topk = 1
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
