% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{prep_parquet_de_all_pairs}
\alias{prep_parquet_de_all_pairs}
\title{INFO -- PLEASE ADD --}
\usage{
prep_parquet_de_all_pairs(
  norm.dat,
  cl,
  cl.bin = NULL,
  cl.bin.size = 100,
  de.param = de_param(),
  method = "fast_limma",
  mc.cores = 1,
  pairs.fn = "pairs.parquet",
  cl.bin.fn = "cl.bin.rda",
  ...
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
