% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{prep_parquet_de_selected_pairs}
\alias{prep_parquet_de_selected_pairs}
\title{INFO -- PLEASE ADD --}
\usage{
prep_parquet_de_selected_pairs(
  norm.dat,
  cl,
  pairs,
  cl.bin = NULL,
  cl.size = NULL,
  de.param = de_parm(),
  method = "fast_limma",
  cl.means = NULL,
  cl.present = NULL,
  cl.sqr.means = NULL,
  use.voom = FALSE,
  counts = NULL,
  mc.cores = 1,
  out.dir = NULL,
  summary.dir = NULL,
  top.n = 500,
  overwrite = FALSE,
  return.df = FALSE,
  return.summary = !is.null(summary.dir)
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
