% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{prepare_train_mmean_list}
\alias{prepare_train_mmean_list}
\title{INFO -- PLEASE ADD --}
\usage{
prepare_train_mmean_list(
  anal_dir =
    "/allen/programs/celltypes/workgroups/rnaseqanalysis/yzizhen/joint_analysis/forebrain_new",
  modality = "10X_cells_v3"
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
