% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNPrepareTaxonomy.R
\name{prepareHANNTaxonomy}
\alias{prepareHANNTaxonomy}
\title{Prepare taxonomy for optimized tree mapping}
\usage{
prepareHANNTaxonomy(AIT.anndata, big.dat = NULL)
}
\arguments{
\item{AIT.anndata}{scrattch.mapping .h5ad object from \code{loadTaxonomy()}}

\item{big.dat}{pre-computed big.data object, @CK can you provide more info on what this is?}
}
\value{
A copy of the anndata object that is also written to disk
}
\description{
This function writes an anndata object in the correct format for all downstream mapping algoriths.
}
