% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTaxonomy.R
\name{prepareTaxonomy}
\alias{prepareTaxonomy}
\title{Prepare taxonomy for optimized tree mapping}
\usage{
prepareTaxonomy(
  count,
  cl,
  cl.df,
  AIT.str,
  lognormal = NULL,
 
    taxonomy.dir = "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/Taxonomies/"
)
}
\arguments{
\item{cl}{assigned cluster}

\item{cl.df}{cluster anno with hierarchy : cluster(cl)/subclass_label/neighborhood/root}

\item{AIT.str}{taxonomy id}

\item{taxonomy.dir}{}

\item{counts}{count\link{gene x cell}}
}
\description{
This function write ...
}
