% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNPrepareTaxonomy.R
\name{rearrange_cl.df}
\alias{rearrange_cl.df}
\title{Rearrange cl.df to internal format labels}
\usage{
rearrange_cl.df(cl.df, cl.hierarchy)
}
\arguments{
\item{counts}{count\link{gene x cell}}

\item{cl}{assigned cluster}
}
\value{
A copy of the anndata object that is also written to disk
}
\description{
This function writes an anndata object in the correct format for all downstream mapping algoriths.
}
\keyword{internal}
