% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{replace_cl_by_label}
\alias{replace_cl_by_label}
\title{INFO -- PLEASE ADD --}
\usage{
replace_cl_by_label(pred, cl.df)
}
\arguments{
\item{cl.df}{to_be_added}

\item{in.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
