% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{replace_class_by_cl}
\alias{replace_class_by_cl}
\title{INFO -- PLEASE ADD --}
\usage{
replace_class_by_cl(pred, cl.df)
}
\arguments{
\item{cl.df}{to_be_added}

\item{in.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
