% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized.R
\name{replace_subclass_by_cl}
\alias{replace_subclass_by_cl}
\title{INFO -- PLEASE ADD --}
\usage{
replace_subclass_by_cl(pred, cl.df)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
