% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_cl}
\alias{resolve_cl}
\title{Title}
\usage{
resolve_cl(
  cl.g,
  cl.dat,
  markers,
  map.dat,
  select.cells,
  p = 0.7,
  low.th = 0.2,
  seed = 42
)
}
\arguments{
\item{cl.g}{}

\item{markers}{}

\item{map.dat}{}

\item{select.cells}{}

\item{p}{}

\item{low.th}{}

\item{seed}{\itemize{
\item random seed for reproducibility
}}

\item{cl.med}{}

\item{dat}{}
}
\value{
mapped.cl output
}
\description{
Title
}
