% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{revert_dend_label}
\alias{revert_dend_label}
\title{Strip extra annotation information from dendrogram}
\usage{
revert_dend_label(dend, value, attribute = "label")
}
\arguments{
\item{dend}{R dendrogram object}

\item{value}{Vector of values pulled from the dendrogram}

\item{attribute}{Which attribute should be overwritten}
}
\value{
R dendrogram object with updated attributes
}
\description{
Strip extra annotation information from dendrogram
}
\keyword{internal}
