% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized.R
\name{run_mapping_on_taxonomy}
\alias{run_mapping_on_taxonomy}
\title{INFO -- PLEASE ADD --}
\usage{
run_mapping_on_taxonomy(
  query.dat,
  Taxonomy = "AIT12.0_mouse",
  prefix = "",
  TaxFN = NA,
  prebuild = FALSE,
  newbuild = FALSE,
  mapping.method = c("flat", "hierarchy"),
  iter = 100,
  mc.cores = 7,
  blocksize = 50000,
  dist.method = "cor",
  topk = 1,
  subsample_pct = 0.9,
  top.n.genes = 15,
  rm.clusters = NA,
  flag.serial = TRUE,
  flag.parallel.tmp = FALSE,
  flag.fuzzy = FALSE
)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
