% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{run_predict_fuzzy_HKNN_cl}
\alias{run_predict_fuzzy_HKNN_cl}
\title{INFO -- PLEASE ADD --}
\usage{
run_predict_fuzzy_HKNN_cl(query, train.list)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
