% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildShiny.R
\name{save_patchseqQC_markers}
\alias{save_patchseqQC_markers}
\title{Save marker genes for patchSeqQC}
\usage{
save_patchseqQC_markers(
  counts,
  metadata,
  subsample = 100,
  subclass.column = "subclass_label",
  class.column = "class_label",
  off.target.types = c("Glia", "glia", "non-neuronal", "Non-neuronal"),
  num.markers = 50,
  shinyFolder = paste0(getwd(), "/")
)
}
\arguments{
\item{counts}{A matrix of counts for the reference data, or character string with a file location of "counts.feather".  If it appears cpm or logCPM matrix is provided, an warning will be thrown.}

\item{metadata}{Data frame of metadata with rows corresponding to cells/nuclei, and either row names or a column called "sample_id" corresponding to cell names. This matrix must include entries for all cells in norm.data.  Could also be a file.  Columns can be numeric, categorical, or factors and must include "subclass.column" and "class.column"}

\item{subsample}{The number of cells to retain per cluster (default = 100).}

\item{subclass.column}{Column name corresponding to the moderate-resolution cell types used for the cell types of interest (default = "subclass_label").}

\item{class.column}{Column name corresponding to the low-resolution cell types used for the off-target cell types (default = "class_label").}

\item{off.target.types}{A character vector of off-target (also known as 'contamination') cell types.  This must include at least one of the cell types found in "class.column" and/or "subclass.column" (both columns are checked)}

\item{num.markers}{The maximum number of markers to calculate per node per direction (default = 50)}

\item{shinyFolder}{= The location to save shiny output (default = current working directory).

Nothing is returned; however an R data object called "QC_markers.RData" is returned with the following variables
markers,
countsQC,
cpmQC,
classBr,
subclassF,
allMarkers,}
}
\description{
This function write a file called \code{QC_markers.RData} that contains all the variables required for applying the patchseq QC algorithm \code{pathseqtools} (which is an more flexible version of the \code{patchSeqQC} algorithm).  This is only used for patch-seq analysis.
}
