% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized.R
\name{select_ancestor_markers}
\alias{select_ancestor_markers}
\title{INFO -- PLEASE ADD --}
\usage{
select_ancestor_markers(nlvl, marker_index, ancestor)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
