% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{select_markers_ds}
\alias{select_markers_ds}
\title{INFO -- PLEASE ADD --}
\usage{
select_markers_ds(de.dir, cl.bin, select.cl = NULL, top.n = 20, mc.cores = 10)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
