% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{select_markers_groups}
\alias{select_markers_groups}
\title{INFO -- PLEASE ADD --}
\usage{
select_markers_groups(
  de.dir,
  cl.group,
  genes,
  cl.bin,
  select.groups = unique(cl.group$group),
  n.markers = 20,
  mc.cores = 1,
  byg = 10,
  ...
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
