% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{select_markers_pair_group_ds}
\alias{select_markers_pair_group_ds}
\title{INFO -- PLEASE ADD --}
\usage{
select_markers_pair_group_ds(
  g1,
  g2,
  de.dir,
  genes,
  cl.bin,
  n.markers = 20,
  select.sign = c("up", "down"),
  max.genes = 50,
  ...
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
