% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{select_markers_pair_group_top_ds}
\alias{select_markers_pair_group_top_ds}
\title{INFO -- PLEASE ADD --}
\usage{
select_markers_pair_group_top_ds(
  g1,
  g2,
  ds,
  genes,
  cl.bin,
  select.sign = "up",
  n.markers = 20,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
