% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_utils.R
\name{select_N_markers_ds}
\alias{select_N_markers_ds}
\title{INFO -- PLEASE ADD --}
\usage{
select_N_markers_ds(
  de.dir,
  select.cl = NULL,
  pair.num = 1,
  add.num = NULL,
  genes,
  cl.bin,
  default.markers = NULL,
  ...
)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
