% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_utils.R
\name{select_pos_markers_ds}
\alias{select_pos_markers_ds}
\title{INFO -- PLEASE ADD --}
\usage{
select_pos_markers_ds(
  de.dir,
  cl,
  select.cl,
  genes,
  cl.bin,
  n.markers = 1,
  mc.cores = 1,
  out.dir = "cl.markers",
  ...
)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
