% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setLeadingString}
\alias{setLeadingString}
\title{Sets the default leading_string for file.path()}
\usage{
setLeadingString(leading_string = NULL)
}
\arguments{
\item{leading_string}{Default (NULL) sets to "\\\\" for Windows and "/" otherwise; or can provide any character vector}
}
\value{
leading_string for use as default in file.path().
}
\description{
Sets the default leading_string for file.path()
}
