% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setPathSeparator}
\alias{setPathSeparator}
\title{Sets the default path_separator for file.path()}
\usage{
setPathSeparator(path_separator = NULL)
}
\arguments{
\item{path_separator}{Default (NULL) sets to .Platform$file.sep or can provide any character vector}
}
\value{
path_separator for use as default in file.path().
}
\description{
Sets the default path_separator for file.path()
}
