% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_mapping.R
\name{taxonomy_mapping}
\alias{taxonomy_mapping}
\title{Cell type annotation and initial QC}
\usage{
taxonomy_mapping(
  GEXRef,
  query.data,
  corr.map = TRUE,
  tree.map = TRUE,
  seurat.map = TRUE,
  label.cols = c("cluster_label", "subclass_label", "class_label"),
  dims = 30,
  k.weight = 15
)
}
\arguments{
\item{GEXRef}{A reference taxonomy object.}

\item{query.data}{A logCPM normalized matrix to be annotated.}

\item{corr.map}{Should correlation mapping be performed?}

\item{tree.map}{Should tree mapping be performed?}

\item{seurat.map}{Should seurat mapping be performed?}

\item{label.cols}{Column names of annotations to map against}

\item{dims}{Number of PCA dimensions for Seurat mapping.}

\item{k.weight}{K neighborhood weight for Seurat mapping.}
}
\value{
Mapping results from all methods.
}
\description{
Perform initial mapping using three methods: Correlation-based, tree-based, and Seurat based, and will calculate some QC metrics.
}
