% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{top_binary_genes}
\alias{top_binary_genes}
\title{Get top genes by beta (binary) score}
\usage{
top_binary_genes(data, cluster.names, gene.count = 2000)
}
\arguments{
\item{data}{A count (or CPM or logCPM) matrix}

\item{cluster.names}{A vector of cluster names in the reference taxonomy.}

\item{gene.count}{The number of top genes to return (Default=2000)}
}
\value{
Boolean vector of cells to keep (TRUE) and cells to remove (FALSE)
}
\description{
Get top genes by beta (binary) score
}
