% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLdivergenceCallAndQC.R
\name{tree_quality_call}
\alias{tree_quality_call}
\title{Calculate tree mapping quality call}
\usage{
tree_quality_call(AIT.anndata, query.mapping)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy object on which buildPatchseqTaxonomy has already been run}

\item{query.mapping}{The output from taxonomy_mapping with both corr.map and tree.map set to TRUE, and with the same mode as currently set in AIT.anndata}
}
\value{
A matrix with "Tree_call" and some other useful columns, one per cell in query.mapping
}
\description{
This function returns the "Good", "I1", "I2", "I3", "PoorQ" calls used to help assess Patch-seq quality.  These scores are based on a combination of KL divergence, correlation to the best matching type, and tree mapping to the top two types.
}
\keyword{external}
