% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap.R
\name{treeMap}
\alias{treeMap}
\title{Tree based mapping}
\usage{
treeMap(GEXRef, query.data)
}
\arguments{
\item{GEXRef}{A reference taxonomy object.}

\item{query.data}{A logCPM normalized matrix to be annotated.}
}
\value{
Tree mapping results as a data.frame.
}
\description{
Tree based mapping
}
