% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_utils.R
\name{update_gene_score_ds}
\alias{update_gene_score_ds}
\title{INFO -- PLEASE ADD --}
\usage{
update_gene_score_ds(
  gene.score,
  ds,
  to.remove,
  cl.bin,
  de = NULL,
  max.num = 1000,
  mc.cores = 20
)
}
\arguments{
\item{cl.df}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
