% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNBuild.R
\name{build_train_list_default}
\alias{build_train_list_default}
\title{INFO -- PLEASE ADD -- @CK please document this function}
\usage{
build_train_list_default(
  pre.train.list = NA,
  query.genes = NA,
  TrainDir,
  TaxDir,
  prefix = "",
  nlevel = 4,
  TaxFN,
  rm.cl = c()
)
}
\arguments{
\item{x}{to_be_added}

\item{key}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD -- @CK please document this function
}
\keyword{internal}
