% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap_optimized_build.R
\name{build_train_list_FB}
\alias{build_train_list_FB}
\title{INFO -- PLEASE ADD --}
\usage{
build_train_list_FB(
  pre.train.list = NA,
  query.genes,
  CLSDIR,
  TaxDir,
  prefix = "",
  nlevel = 4,
  TaxFN,
  rm.cl = c()
)
}
\arguments{
\item{key}{}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
