% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNBuild.R
\name{build_train_list_on_taxonomy}
\alias{build_train_list_on_taxonomy}
\title{Starting point for optimized tree mapping}
\usage{
build_train_list_on_taxonomy(
  TaxFN = NA,
  Taxonomy,
  TaxHome = "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/Taxonomies/",
  pre.train.list = NA,
  query.genes = NA,
  prefix = "",
  mapping.method = c("flat", "one-step", "hierarchy"),
  nlevel = 2,
  prebuild = FALSE,
  newbuild = FALSE,
  mc.cores = 10,
  div_thr = 3,
  subsample_pct = 0.9,
  top.n.genes = 15,
  n.group.genes = 3000,
  rm.cl = c()
)
}
\arguments{
\item{TaxFN}{to_be_added}

\item{Taxonomy}{to_be_added}

\item{pre.train.list}{to_be_added}

\item{query.genes}{to_be_added}

\item{prefix}{to_be_added}

\item{mapping.method}{to_be_added}

\item{prebuild}{to_be_added}

\item{newbuild}{to_be_added}

\item{mc.cores}{to_be_added}

\item{div_thr}{to_be_added}

\item{subsample_pct}{to_be_added}

\item{top.n.genes}{to_be_added}

\item{n.group.genes}{to_be_added}

\item{rm.cl}{to_be_added}
}
\value{
Mapping results
}
\description{
Starting point for optimized tree mapping
}
