% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildShiny.R
\name{buildReference}
\alias{buildReference}
\title{Starting from a Seurat object this function builds the minimum files required for Shiny}
\usage{
buildReference(
  counts,
  meta.data,
  feature.set,
  umap.coords,
  shinyFolder,
  subsample = 2000,
  reorder.dendrogram = FALSE
)
}
\arguments{
\item{counts}{A count matrix in sparse format: dgCMatrix.}

\item{meta.data}{Meta.data corresponding to count matrix. Rownames must be equal to colnames of counts.}

\item{feature.set}{Set of feature used to calculate dendrogram. Typically highly variable and/or marker genes.}

\item{umap.coords}{Dimensionality reduction coordiant data.frame with 2 columns. Rownames must be equal to colnames of counts.}

\item{shinyFolder}{The location to save Shiny objects, e.g. "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/10x_seq/NHP_BG_20220104/"}

\item{subsample}{The number of cells to retain per cluster}

\item{reorder.dendrogram}{Should dendogram attempt to match a preset order? (Default = FALSE).  If TRUE, the dendrogram attempts to match the celltype factor order as closely as possible (if celltype is a character vector rather than a factor, this will sort clusters alphabetically, which is not ideal).}
}
\value{

}
\description{
Starting from a Seurat object this function builds the minimum files required for Shiny
}
