% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildShiny.R
\name{buildReferenceFolder}
\alias{buildReferenceFolder}
\title{Starting from a Seurat object this function builds the minimum files required for Shiny}
\usage{
buildReferenceFolder(
  seurat.obj,
  shinyFolder,
  subsample = 2000,
  feature.set = NULL,
  reorder.dendrogram = FALSE,
  save.normalized.data = FALSE
)
}
\arguments{
\item{seurat.obj}{A Seurat object as specified in the notes, or as output from "createSeuratObjectForReferenceFolder"}

\item{shinyFolder}{The location to save Shiny objects, e.g. "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/10x_seq/NHP_BG_20220104/"}

\item{subsample}{The number of cells to retain per cluster}

\item{feature.set}{An optional user supplied gene set otherwise Seurat obects variable features are used}

\item{reorder.dendrogram}{Should dendogram attempt to match a preset order? (Default = FALSE).  If TRUE, the dendrogram attempts to match the celltype factor order as closely as possible (if celltype is a character vector rather than a factor, this will sort clusters alphabetically, which is not ideal).}

\item{save.normalized.data}{Should normalized data (TRUE="data" slot in Seurat object) or raw count matrix (FALSE="counts" slot in Seurat object; default) be saved?
\subsection{NOTES}{

Seurat object must have the following:
Expr. data is gathered from: GetAssayData(seurat.obj, "data") and "counts"
Either VariableFeatures(seurat.obj) or feature.set parameter must be supplied.
seurat.obj@meta.data$cluster contains clustering results, used both for subsampling and dendrogram labels
seurat.obj@meta.data$sample_id is set via colnames(seurat.obj)
UMAP/tSNE is pulled from: FetchData(seurat.obj, vars=c("UMAP_1","UMAP_2")), simply run Seurats UMAP function or add a new dimensionSlot with the UMAP_ name.
}}
}
\value{

}
\description{
Starting from a Seurat object this function builds the minimum files required for Shiny
}
