% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN.R
\name{call_ANN_0}
\alias{call_ANN_0}
\title{INFO -- PLEASE ADD --}
\usage{
call_ANN_0(index.bs, qdat, prev, children, gg)
}
\arguments{
\item{index.bs}{to_be_added}

\item{qdat}{to_be_added}

\item{prev}{to_be_added}

\item{children}{to_be_added}

\item{gg}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
