% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrMap.R
\name{cor}
\alias{cor}
\title{Correlation based cell type mapping wrapper}
\usage{
cor(...)
}
\arguments{
\item{input_logcpm}{Inputted log2(CPM+1) values.}

\item{reference_medians}{Median expression levels (log2(CPM+1)) of reference cell types.}

\item{nGenes}{Number of high variance genes to include.}
}
\value{
Correlation mapping results
}
\description{
Correlation based cell type mapping wrapper
}
\keyword{internal}
