% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{cor_ancestor_markers}
\alias{cor_ancestor_markers}
\title{INFO -- PLEASE ADD --}
\usage{
cor_ancestor_markers(qdat, assigned.df, train.cl.dat, ancestor_markers)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
