% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrMap.R
\name{corrMap}
\alias{corrMap}
\title{Correlation based mapping}
\usage{
corrMap(AIT.anndata, query.data)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}

\item{query.data}{A logCPM normalized matrix to be annotated.}
}
\value{
Correlation mapping results as a data.frame.
}
\description{
Correlation based mapping
}
