% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{create_initial_df}
\alias{create_initial_df}
\title{INFO -- PLEASE ADD --}
\usage{
create_initial_df(in_sample_name, in_best.cl = NA)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
