% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{create_initial_membership}
\alias{create_initial_membership}
\title{INFO -- PLEASE ADD --}
\usage{
create_initial_membership(in_sample_name, ancestor)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
