% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{get_de_genes}
\alias{get_de_genes}
\title{INFO -- PLEASE ADD --}
\usage{
get_de_genes(ds, cl.bin, cl1, cl2)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
