% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{get_knn}
\alias{get_knn}
\title{Get KNN}
\usage{
get_knn(
  dat,
  ref.dat,
  k,
  method = "cor",
  dim = NULL,
  index = NULL,
  build.index = FALSE,
  transposed = TRUE,
  return.distance = FALSE
)
}
\arguments{
\item{dat}{to_be_added}

\item{ref.dat}{to_be_added}

\item{k}{to_be_added}

\item{method}{to_be_added}

\item{dim}{to_be_added}
}
\value{
to_be_added
}
\description{
Get KNN
}
\keyword{internal}
