% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{get_last_field}
\alias{get_last_field}
\title{INFO -- PLEASE ADD --}
\usage{
get_last_field(x, key)
}
\arguments{
\item{x}{to_be_added}

\item{key}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
