% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxonomy.R
\name{loadTaxonomy}
\alias{loadTaxonomy}
\title{Read in a reference data set in Allen taxonomy format}
\usage{
loadTaxonomy(
  refFolder,
  sample_id = "sample_id",
  hGenes = NULL,
  gene_id = "gene"
)
}
\arguments{
\item{refFolder}{Directory containing the Shiny taxonomy.}

\item{sample_id}{Field in reference taxonomy that defines the sample_id.}

\item{hGenes}{User supplied variable gene vector.  If not provided, then all genes are used.}

\item{gene_id}{Field in counts.feather that defines the gene_id.}
}
\value{
Organized reference object ready for mapping against.
}
\description{
Read in a reference data set in Allen taxonomy format
}
