% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxonomy.R
\name{loadTaxonomy}
\alias{loadTaxonomy}
\title{Read in a reference data set in Allen taxonomy format}
\usage{
loadTaxonomy(
  taxonomyDir,
  anndata_file = "AI_taxonomy.h5ad",
  sample_id = "sample_id",
  hGenes = NULL,
  gene_id = "gene",
  force = FALSE
)
}
\arguments{
\item{taxonomyDir}{Directory containing the Shiny taxonomy -OR- a direct h5ad file name.}

\item{anndata_file}{File name of the anndata object to be loaded.}

\item{sample_id}{Field in reference taxonomy that defines the sample_id.}

\item{hGenes}{User supplied variable gene vector.  If not provided, then all genes are used.}

\item{gene_id}{Field in counts.feather that defines the gene_id.}

\item{force}{Force rebuild the anndata object for the taxonomy.}

\item{patchseq}{Should the patchseq variant of this taxonomy be loaded.}
}
\value{
Organized reference object ready for mapping against.
}
\description{
Read in a reference data set in Allen taxonomy format
}
