% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{map_cells_knn}
\alias{map_cells_knn}
\title{INFO -- PLEASE ADD --}
\usage{
map_cells_knn(
  topk = 1,
  test.dat,
  cl.dat,
  train.index = NULL,
  method = c("Annoy.Cosine", "cor"),
  batch.size = 5000,
  mc.cores = 1
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
