% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_dend}
\alias{map_dend}
\title{map_dend}
\usage{
map_dend(
  dend,
  cl.dat,
  map.dat,
  select.cells = colnames(map.dat),
  p = 0.8,
  low.th = 0.1,
  default.markers = NULL,
  seed = 42
)
}
\arguments{
\item{dend}{A dendrogram in R format}

\item{cl.dat}{gene by cell type matrix (I think?)}

\item{map.dat}{normalized data of the MAPPING data set.}

\item{select.cells}{names of cells to map (e.g., the column names of the cell x gene matrix)}

\item{p}{proportion of marker genes to include in each iteration of the mapping algorithm.}

\item{low.th}{the minimum difference in Pearson correlation required to decide on which branch}

\item{default.markers}{What genes to include in every bootstrap run (default is none)}

\item{seed}{= random seed}

\item{cl}{A cluster factor object to compare to a reference}
}
\value{
tree mapping to the dendrogram table (cells x nodes with values as probabilities)
}
\description{
map_dend
}
\keyword{internal}
