% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_dend_membership}
\alias{map_dend_membership}
\title{map_dend_membership}
\usage{
map_dend_membership(
  dend,
  cl.dat,
  map.dat,
  map.cells,
  mc.cores = 10,
  bs.num = 100,
  seed = 42,
  ...
)
}
\arguments{
\item{dend}{R dendrogram in a specific format}

\item{cl.dat}{gene by cell type matrix (I think?)}

\item{map.dat}{normalized data of the MAPPING data set.}

\item{map.cells}{names of cells to map (e.g., the column names of the cell x gene matrix)}

\item{mc.cores}{number of cores to run the mapping on}

\item{bs.num}{Number of bootstrapping runs to calculate the membership from (default = 100)}

\item{seed}{= random seed}

\item{...}{other variables to pass to map_dend}
}
\value{
membership table
}
\description{
map_dend_membership
}
\keyword{internal}
