% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN.R
\name{mapping_on_taxonomy}
\alias{mapping_on_taxonomy}
\title{INFO -- PLEASE ADD --}
\usage{
mapping_on_taxonomy(
  query.dat,
  train.list,
  blocksize = 5000,
  mc.cores = 5,
  method = "cor",
  iter = 100,
  topk = 1,
  flag.serial = TRUE,
  flag.parallel.tmp = FALSE,
  flag.fuzzy = FALSE
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
