% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN_prepareTaxonomy.R
\name{prepareTaxonomy}
\alias{prepareTaxonomy}
\title{Prepare taxonomy for optimized tree mapping}
\usage{
prepareTaxonomy(
  count,
  cl,
  cl.df,
  AIT.str,
  lognormal = NULL,
 
    taxonomy.dir = "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/Taxonomies/"
)
}
\arguments{
\item{cl}{assigned cluster}

\item{cl.df}{cluster anno with hierarchy : cluster(cl)/subclass_label/neighborhood/root}

\item{AIT.str}{taxonomy id}

\item{lognormal}{a logCPM cell x gene matrix.  If counts provided, logCPM is calculated.}

\item{taxonomy.dir}{Output directly to write h5ad file}

\item{counts}{count\link{gene x cell}}
}
\value{
A copy of the anndata object that is also written to disk
}
\description{
This function writes an anndata object in the correct format for all downstream mapping algoriths.
}
