% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN_prepareTaxonomy.R
\name{prepareTaxonomy}
\alias{prepareTaxonomy}
\title{Prepare taxonomy for optimized tree mapping}
\usage{
prepareTaxonomy(
  count,
  cl,
  cl.df.users,
  cl.hierarchy,
  AIT.str,
  taxonomy.dir = "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/Taxonomies/",
  h5ad_out = FALSE
)
}
\arguments{
\item{cl}{assigned cluster}

\item{cl.df.users}{cluster annotation}

\item{cl.hierarchy}{: hierarhcy in clusters : cluster(cl)/subclass_label/neighborhood/root}

\item{AIT.str}{taxonomy id}

\item{taxonomy.dir}{Output directly to write h5ad file}

\item{counts}{count\link{gene x cell}}
}
\value{
A copy of the anndata object that is also written to disk
}
\description{
This function writes an anndata object in the correct format for all downstream mapping algoriths.
}
