% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_cl}
\alias{resolve_cl}
\title{resolve_cl}
\usage{
resolve_cl(
  cl.g,
  cl.dat,
  markers,
  map.dat,
  select.cells,
  p = 0.8,
  low.th = 0.1,
  seed = 42
)
}
\arguments{
\item{cl.g}{Cluster labels in some format}

\item{markers}{Genes to use as markers for this function}

\item{map.dat}{normalized data of the MAPPING data set.}

\item{select.cells}{names of cells to map (e.g., the column names of the cell x gene matrix)}

\item{p}{proportion of marker genes to include in each iteration of the mapping algorithm.}

\item{low.th}{the minimum difference in Pearson correlation required to decide on which branch}

\item{seed}{\itemize{
\item random seed for reproducibility
}}

\item{cl.med}{Cluster medians}
}
\value{
mapped.cl output
}
\description{
resolve_cl
}
\keyword{internal}
