% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN_prepareTaxonomy.R
\name{run_prepareTaxonomy}
\alias{run_prepareTaxonomy}
\title{run_prepareTaxonomy : call prepareTaxonomy() from h5ad file}
\usage{
run_prepareTaxonomy(h5adFN)
}
\arguments{
\item{h5adFN}{filename of anndata with uns[\link{'HANN'}]}
}
\description{
run_prepareTaxonomy : call prepareTaxonomy() from h5ad file
}
