% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANN.R
\name{select_ancestor_markers}
\alias{select_ancestor_markers}
\title{INFO -- PLEASE ADD --}
\usage{
select_ancestor_markers(nlvl, marker_index, ancestor)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
