% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{select_markers_groups_top_ds}
\alias{select_markers_groups_top_ds}
\title{INFO -- PLEASE ADD --}
\usage{
select_markers_groups_top_ds(
  ds,
  cl.group,
  select.groups = names(cl.group),
  n.markers = 3,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
