% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{select_markers_pair_direction_ds}
\alias{select_markers_pair_direction_ds}
\title{INFO -- PLEASE ADD --}
\usage{
select_markers_pair_direction_ds(
  de.dir,
  add.num,
  genes,
  cl.bin,
  de = NULL,
  mc.cores = mc.cores,
  max.genes = 1000,
  byg = 10,
  ...
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
