% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{select_top_pos_markers_ds}
\alias{select_top_pos_markers_ds}
\title{INFO -- PLEASE ADD --}
\usage{
select_top_pos_markers_ds(
  ds,
  cl,
  select.cl,
  genes,
  cl.bin,
  n.markers = 3,
  mc.cores = 10,
  ...
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
