% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratMap.R
\name{seuratMap}
\alias{seuratMap}
\title{Seurat based mapping}
\usage{
seuratMap(GEXRef, query.data, dims = 30, k.weight = 5)
}
\arguments{
\item{GEXRef}{A reference taxonomy object.}

\item{query.data}{A logCPM normalized matrix to be annotated.}
}
\value{
Seurat mapping results as a data.frame.
}
\description{
Seurat based mapping
}
