% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomyMapping.R
\name{taxonomy_mapping}
\alias{taxonomy_mapping}
\title{Cell type annotation and initial QC}
\usage{
taxonomy_mapping(
  AIT.anndata,
  query.data,
  corr.map = TRUE,
  tree.map = TRUE,
  seurat.map = TRUE,
  label.cols = c("cluster_label", "subclass_label", "class_label")
)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy object.}

\item{query.data}{A logCPM normalized matrix to be annotated.}

\item{corr.map}{Should correlation mapping be performed?}

\item{tree.map}{Should tree mapping be performed?}

\item{seurat.map}{Should seurat mapping be performed?}

\item{label.cols}{Column names of annotations to map against.  Note that this only works for metadata that represent clusters or groups of clusters (e.g., subclass, supertype, neighborhood, class)}
}
\value{
Mapping results from all methods.
}
\description{
Perform initial mapping using three methods: Correlation-based, tree-based, and Seurat based, and will calculate some QC metrics.
}
