% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap.R
\name{treeMap}
\alias{treeMap}
\title{Tree based mapping}
\usage{
treeMap(AIT.anndata, query.data)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}

\item{query.data}{A logCPM normalized matrix to be annotated.}
}
\value{
Tree mapping results as a data.frame.
}
\description{
Tree based mapping
}
