% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNUtils.R
\name{update_gene_score_ds}
\alias{update_gene_score_ds}
\title{INFO -- PLEASE ADD --}
\usage{
update_gene_score_ds(
  gene.score,
  ds,
  to.remove,
  cl.bin,
  de = NULL,
  max.num = 1000,
  mc.cores = 20
)
}
\arguments{
\item{in.df}{to_be_added}

\item{cl.df}{to_be_added}
}
\value{
???
}
\description{
INFO -- PLEASE ADD --
}
\keyword{internal}
